rename helpers.test.ts -> testHelpers.ts
jest fails if there are test files that match the test pattern that do not define any tests. It is possible to configure around this, but it's simpler to just rename these files.
