+++ shared/src/api/extension/types/position.test.ts
@@ -1,5 +1,5 @@
-import { assertToJSON } from './common.test'
 import { Position } from './position'
+import { assertToJSON } from './testHelpers'
 
 describe('Position', () => {
     test('constructs', () => {
